requirements.txt
flask
joblib
pandas
numpy
scikit-learn==1.6.1
mlflow
requests
streamlit
plotly
scipy
gunicorn
python-dotenv
